restructure project and create a test
